requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.0
Flask-Migrate==4.0.3
python-dotenv==1.0.0
requests==2.28.2  
Flask-Cors==3.1.1